requirements.txt
Flask==3.1.0
Faker==37.1.0
psycopg2-binary==2.9.10
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-system-metrics